Add image query functionality to leaflet/mapview map.

Add image query functionality to leaflet/mapview map.

addImageQuery(map, x, band = 1, group = NULL, layerId = NULL,
  project = TRUE, type = c("mousemove", "click"), digits,
  position = "topright", prefix = "Layer", ...)

Arguments

map
the map with the RasterLayer to be queried.

x
the RasterLayer that is to be queried.

band
for stars layers, the band number to be queried.

group
the group of the RasterLayer to be queried.

layerId
the layerId of the RasterLayer to be queried. Needs to be the same as supplied in addRasterImage or link{addStrasImage}.

project
whether to project the RasterLayer to conform with leaflets expected crs. Defaults to TRUE and things are likely to go haywire if set to FALSE.

type
whether query should occur on 'mousemove' or 'click'. Defaults to 'mousemove'.

digits
the number of digits to be shown in the display field.

position
where to place the display field. Default is 'topright'.

prefix
a character string to be shown as prefix for the layerId.

...
currently not used.

Details

This function is deprecated. Please use leafem::addImageQuery instead.