Add a tiled raster image

Create tiled image data and add to a leaflet or mapview map. The image data can be in the form of a character vector pointing to a file on disk such as .png, .jpg, .bmp, .tif files or a RasterLayer from the raster package.

NOTE:
In case you want to save your map with tiled imagery using htmlwidgets::saveWidget or mapview::mapshot make sure you set libdir = "lib" as otherwise tiles won't render in your .html.

addTiledRasterImage(map, x, minzoom = 0, maxzoom,
  color = viridisLite::inferno, at, na.color = "#BEBEBE")

addTiledImage(map, x, minzoom = 0, maxzoom)

Arguments

map
a leaflet map

x
for addTiledRasterImage a RasterLayer, for addTiledImage a file path to the image to be tiled.

minzoom
minimum zoom of the map

maxzoom
maximum zoom of the map. This will be set as the maximum zoom during tile generation. If missing, it is calculated based on image dimensions.

color
the color palette to be used

at
optionally specify breaks

na.color
color for NA values

Details

add non-projected/non-geographical tiled images to a non-projected leaflet map for better performance. The map should be set up with leafletCRS(crsClass = "L.CRS.Simple") - see example below.

The use of these functions requires some system libraries to be installed. First and foremost, GDAL needs to be available on your system. Furthermore, you will need to have the python-gdal bindings. Installing GDAL is pretty straight forward for most systems and chances are you have it already, given that you were able to install mapview. The gdal-python bindings require some extra work.

If you're useing a debian linux system

sudo apt-get install python-gdal

should be sufficient.

On windows (as usual) things are a more involved.
The following describes how to get things set up to get these functions working under windows:

Download and install Python from https://www.python.org/ftp/python, e.g. python-2.7.9.amd64.msi for Python 2.7 64-bit (assumed in the following)

From http://www.gisinternals.com/release.php, download and install
1. GDAL core, e.g. gdal-203-1911-x64-core.msi,
2. then GDAL Python bindings, e.g. GDAL-2.3.0.win-amd64-py2.7.msi

Add required GDAL paths to system variables:
1. PATH: C:\Program Files (x86)\GDAL
2. GDAL_DATA: C:\Program Files (x86)\GDAL\gdal-data
3. GDAL_DRIVER_PATH: C:\Program Files (x86)\GDAL\gdalplugins

Download and install Microsoft Visual C++ Compiler for Python 2.7 from https://www.microsoft.com/en-us/download/confirmation.aspx?id=44266 (required for installation of numpy package)

Download and extract numpy (e.g. numpy-1.14.3.zip from https://files.pythonhosted.org/packages/b0/2b/497c2bb7c660b2606d4a96e2035e92554429e139c6c71cdff67af66b58d2/numpy-1.14.3.zip). On the command line, navigate to decompressed folder and run

>>> python setup.py install

To verify, run the following import statements in the Python shell on the command line:

>>> import gdal
>>> import ogr
>>> import osr
>>> import gdalnumeric
>>> import gdalconst

See also

https://pypi.org/project/GDAL/ and

https://sandbox.idre.ucla.edu/sandbox/tutorials/installing-gdal-for-windows

for further details.

On a mac you're on your own for now (we don't have testing opportunities). Though we would appreciate a PR if you feel confident you have a solution for installing all needed dependencies/prerequisites.

Examples

# NOT RUN {
library(leaflet)
library(raster)
library(mapview)

# generate a large RasterLayer
rst = raster(nrows = 5000, ncols = 10000,
             xmn = 0, xmx = 10000,
             ymn = -5000, ymx = 0)
rst[] = runif(ncell(rst))

# generate leaflet map with simple crs
map = leaflet(
  options = leaflet::leafletOptions(
    crs = leafletCRS(
      crsClass = "L.CRS.Simple"
    )
  )
)

options(viewer = NULL)

map %>%
  addTiledRasterImage(rst)
# }