View two or more (possibly synchronised) mapview or leaflet maps

These functions are deprecated. Please use leafsync::sync and leafsync::latticeView instead.

latticeView(..., ncol = 2, sync = "none", sync.cursor = FALSE,
  no.initial.sync = TRUE)

latticeview(...)

sync(..., ncol = 2, sync = "all", sync.cursor = TRUE,
  no.initial.sync = TRUE)

Arguments

...
any number of mapview or leaflet objects or a list thereof

ncol
how many columns should be plotted

sync
whether to synchronise zoom and pan for certain elements. Possible values are "all" (default) to sync all maps, "none" to disable synchronisation or a list of panel numbers, e.g. list(c(1, 3), c(2, 4)) will synchronise panels 1 & 3 and panels 2 & 4. Panels are drawn from top right to bottom left.

sync.cursor
whether to show cursor position in synced panels (default TRUE).

no.initial.sync
whether to sync the initial view (default TRUE).

Functions

latticeview: alias for ease of typing

sync: convenience function for syncing maps