mapview: Interactive Viewing of Spatial Data in R

Quickly and conveniently create interactive visualisations of spatial data with or without background maps. Attributes of displayed features are fully queryable via pop-up windows. Additional functionality includes methods to visualise true- and false-color raster images and bounding boxes.

Details

The package provides functionality to view spatial objects interactively. The intention is to provide interactivity for easy and quick visualization during spatial data analysis. It is not intended for fine-tuned presentation quality map production.

See also

Author

Maintainer: Tim Appelhans tim.appelhans@gmail.com

Authors:

Florian Detsch fdetsch@web.de

Christoph Reudenbach reudenbach@geo.uni-marburg.de

Stefan Woellauer stephan.woellauer@geo.uni-marburg.de

Other contributors:

Spaska Forteva spaska.forteva@geo.uni-marburg.de [contributor]

Thomas Nauss nauss@staff.uni-marburg.de [contributor]

Edzer Pebesma [contributor]

Kenton Russell [contributor]

Michael Sumner [contributor]

Jochen Darley Debugger@jedimasters.de [contributor]

Pierre Roudier [contributor]

Patrick Schratz [contributor]

Environmental Informatics Marburg [contributor]

Lorenzo Busetto [contributor]