mapview + mapview adds data from the second map to the first

mapview + mapview adds data from the second map to the first

mapview + data adds spatial data (raster*, sf*, sp*) to a mapview map

mapview + NULL returns the LHS map

[...]

mapview | mapview provides a slider in the middle to compare two maps.

mapview | NULL returns the LHS map

NULL | mapview returns the RHS map

# S4 method for mapview,mapview
+(e1, e2)

# S4 method for mapview,ANY
+(e1, e2)

# S4 method for mapview,`NULL`
+(e1, e2)

# S4 method for mapview,character
+(e1, e2)

# S4 method for mapview,`NULL`
|(e1, e2)

# S4 method for `NULL`,mapview
|(e1, e2)

Arguments

e1

a leaflet or mapview map, or NULL.

e2

a leaflet or mapview map, or NULL.

Examples

  m1 <- mapView(franconia, col.regions = "red")
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019
  m2 <- mapView(breweries)
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019

  ### add two mapview objects
  m1 + m2

  ### add layers to a mapview object
  if (interactive()) {
    library(plainview)
    m1 + breweries + plainview::poppendorf[[4]]
  }

  m1 <- mapView(franconia, col.regions = "red")
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019
  m2 <- mapView(breweries)
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019

  ### add two mapview objects
  m1 | m2
#> Loading required namespace: leaflet.extras2