mapview + mapview adds data from the second map to the first

mapview + mapview adds data from the second map to the first

mapview + data adds spatial data (raster*, sf*, sp*) to a mapview map

[...]

# S4 method for mapview,mapview
+(e1, e2)

# S4 method for mapview,ANY
+(e1, e2)

# S4 method for mapview,character
+(e1, e2)

Arguments

e1
a leaflet or mapview map to which e2 should be added.

e2
a (spatial) object to be added or a mapview object from which the objects should be added to e1.

Examples

not_run({
  ### raster data ###
  library(sp)
  library(raster)

  m1 <- mapView(poppendorf[[5]])

  ### point vector data ###
  m2 <- mapView(breweries91)

  ### add two mapview objects
  m1 + m2 # final zoom level based on m2
  '+'(m2, m1) # final zoom level based on m1
})