Create HTML strings for popups

These functions are deprecated. Please use leafpop::popupTable, leafpop::popupImage and leafpop::popupGraph instead.

popupTable(x, zcol, row.numbers = TRUE, feature.id = TRUE)

popupImage(img, src = c("local", "remote"), embed = FALSE, ...)

popupGraph(graphs, type = c("png", "svg", "html"), width = 300,
  height = 300, ...)

Arguments

x
A Spatial* object.

zcol
numeric or character vector indicating the columns included in the output popup table. If missing, all columns are displayed.

row.numbers
logical whether to include row numbers in the popup table.

feature.id
logical whether to add 'Feature ID' entry to popup table.

img
A character vector of file path(s) or web-URL(s) to any sort of image file(s).

src
Whether the source is "local" (i.e. valid file path(s)) or "remote" (i.e. valid URL(s)).

embed
whether to embed the (local) images in the popup html as base64 ecoded. Set this to TRUE if you want to save and share your map, unless you want render many images, then set to FALSE and make sure to copy ../graphs when copying the map to a different location.

...
further arguments passed on to underlying methods such as height and width.

graphs
A list of figures associated with x.

type
Output filetype, one of "png" (default), "svg" or "html".

width
popup width in pixels.

height
popup height in pixels.