Convert a vector/matrix of coordinates to JSON format

Similar to toJSON from jsonlite, this function takes a set of coordinates as input and converts them to proper JSON format. Note that the function is powered by Rcpp which makes it a convenient alternative to existing methods when it comes to processing big datasets.

# S4 method for numeric
coords2JSON(x)

# S4 method for character
coords2JSON(x, xy = c(1, 2))

# S4 method for matrix
coords2JSON(x, xy = c(1, 2))

Arguments

x
A 'numeric' vector with a single pair of coordinates or a matrix with multiple pairs of input coordinates, typically projected in EPSG:4326 (http://spatialreference.org/ref/epsg/wgs-84/).

xy
An 'integer' vector specifying the coordinate columns.

Value

A single 'character' object in JSON format.

Examples

crd <- matrix(ncol = 3, nrow = 12)

# x-coordinates
set.seed(10)
crd[, 1] <- rnorm(nrow(crd), 10, 3)

# y-coordinates
set.seed(10)
crd[, 2] <- rnorm(nrow(crd), 50, 3)

# additional data
crd[, 3] <- month.abb

# reformat a single pair of coordinates
coords2JSON(crd[1, ])

#> [1] "[10.0562385,50.0562385,\"Jan\"]"

# reformat multiple pairs of coordinates at once
coords2JSON(crd)

#> [1] "[[10.0562385,50.0562385,\"Jan\"],[9.4472424,49.4472424,\"Feb\"],[5.8860084,45.8860084,\"Mar\"],[8.2024969,48.2024969,\"Apr\"],[10.8836354,50.8836354,\"May\"],[11.1693829,51.1693829,\"Jun\"],[6.3757715,46.3757715,\"Jul\"],[8.9089719,48.9089719,\"Aug\"],[5.119982,45.119982,\"Sep\"],[9.2305648,49.2305648,\"Oct\"],[13.3053385,53.3053385,\"Nov\"],[12.2673445,52.2673445,\"Dec\"]]"