Defunct functions in mapview

These functions have been removed from package mapview. See below for information on which package they have been moved to.

Details

cubeview: This function is defunct, and has been migrated to package 'cubeview'.

cubeView: This function is defunct, and has been migrated to package 'cubeview'.

cubeViewOutput: This function is defunct, and has been migrated to package 'cubeview'.

renderCubeView: This function is defunct, and has been migrated to package 'cubeview'.

slideview: This function is defunct, and has been migrated to package 'slideview'.

slideView: This function is defunct, and has been migrated to package 'slideview'.

slideViewOutput: This function is defunct, and has been migrated to package 'slideview'.

renderslideView: This function is defunct, and has been migrated to package 'slideview'.

latticeView: This function is defunct, and has been migrated to package 'leafsync'.

sync: This function is defunct, and has been migrated to package 'leafsync'.

plainview: This function is defunct, and has been migrated to package 'plainview'.

plainView: This function is defunct, and has been migrated to package 'plainview'.

popupTable: This function is defunct, and has been migrated to package 'leafpop'.

popupImage: This function is defunct, and has been migrated to package 'leafpop'.

popupGraph: This function is defunct, and has been migrated to package 'leafpop'.

addFeatures: This function is defunct, and has been migrated to package 'leafem'.

garnishMap: This function is defunct, and has been migrated to package 'leafem'.

addHomeButton: This function is defunct, and has been migrated to package 'leafem'.

removeHomeButton: This function is defunct, and has been migrated to package 'leafem'.

addImageQuery: This function is defunct, and has been migrated to package 'leafem'.

addLogo: This function is defunct, and has been migrated to package 'leafem'.

addMouseCoordinates: This function is defunct, and has been migrated to package 'leafem'.

removeMouseCoordinates: This function is defunct, and has been migrated to package 'leafem'.

addStaticLabels: This function is defunct, and has been migrated to package 'leafem'.

addExtent: This function is defunct, and has been migrated to package 'leafem'.

addStarsImage: This function is defunct, and has been migrated to package 'leafem'.