Create HTML strings for popups

Create HTML strings for popup tables used as input for mapview or leaflet. This optionally allows the user to include only a subset of feature attributes.

Create HTML strings for popup images used as input for mapview or leaflet.

Create HTML strings for popup graphs used as input for mapview or leaflet.

popupTable(x, zcol, row.numbers = TRUE)

popupImage(img, src = c("local", "remote"), ...)

popupGraph(graphs, type = c("png", "svg", "html"), width = 300,
  height = 300, ...)

Arguments

x
A Spatial* object.

zcol
numeric or character vector indicating the columns included in the output popup table. If missing, all columns are displayed.

row.numbers
logical whether to include row numbers in the popup table.

img
A character vector of file path(s) or web-URL(s) to any sort of image file(s).

src
Whether the source is "local" (i.e. valid file path(s)) or "remote" (i.e. valid URL(s)).

...
further arguments passed on to underlying methods such as height and width.

graphs
A list of figures associated with x.

type
Output filetype, one of "png" (default), "svg" or "html".

width
popup width in pixels.

height
popup height in pixels.

Value

A list of HTML strings required to create feature popup tables.

A list of HTML strings required to create popup graphs.

A list of HTML strings required to create popup graphs.

Details

Type svg uses native svg encoding via readLines. height and width are set via ... and passed on to svg
Type png embeds via "<img src = ...". height and width are set via ... and passed on to png
Type html embeds via "<iframe src = ...". height and width are set directly in pixels.

Examples

library(leaflet)

## include columns 1 and 2 only
mapview(franconia, popup = popupTable(franconia, zcol = 1:2))
mapview(breweries, zcol = "founded", legend = TRUE,
        popup = popupTable(breweries, zcol = c("founded", "village")))
leaflet() %>% addCircleMarkers(data = breweries)
leaflet() %>% addCircleMarkers(data = breweries,
                               popup = popupTable(breweries))
# NOT RUN {
## remote images -----
### one image
library(sf)

pnt = st_as_sf(data.frame(x = 174.764474, y = -36.877245),
                coords = c("x", "y"),
                crs = 4326)

img = "http://bit.ly/1TVwRiR"

mapview(pnt, popup = popupImage(img, src = "remote"))

### multiple file (types)
library(sp)
images = c(img,
            "https://upload.wikimedia.org/wikipedia/commons/9/91/Octicons-mark-github.svg",
            "https://www.r-project.org/logo/Rlogo.png",
            "https://upload.wikimedia.org/wikipedia/commons/d/d6/MeanMonthlyP.gif")

pt4 = data.frame(x = jitter(rep(174.764474, 4), factor = 0.01),
                  y = jitter(rep(-36.877245, 4), factor = 0.01))
coordinates(pt4) = ~ x + y
proj4string(pt4) = "+init=epsg:4326"

mapview(pt4, popup = popupImage(images)) # NOTE the gif animation

## local images -----
pnt = st_as_sf(data.frame(x = 174.764474, y = -36.877245),
                coords = c("x", "y"), crs = 4326)
img = system.file("img","Rlogo.png",package="png")
mapview(pnt, popup = popupImage(img))
# }

# NOT RUN {
### example: svg -----

library(sp)

data(meuse)
coordinates(meuse) = ~ x + y
proj4string(meuse) = CRS("+init=epsg:28992")

## create plots with points colored according to feature id
library(lattice)
p = xyplot(copper ~ cadmium, data = meuse@data, col = "grey")
p = mget(rep("p", length(meuse)))

clr = rep("grey", length(meuse))
p = lapply(1:length(p), function(i) {
  clr[i] = "red"
  update(p[[i]], col = clr)
})

mapview(meuse, popup = popupGraph(p, type = "svg"))

### example: png -----
pt = data.frame(x = 174.764474, y = -36.877245)

coordinates(pt) = ~ x + y
proj4string(pt) = "+init=epsg:4326"

p2 = levelplot(t(volcano), col.regions = terrain.colors(100))

mapview(pt, popup = popupGraph(p2, width = 300, height = 400))

### example: html -----
mapview(breweries[1, ], map.types = "Esri.WorldImagery",
        popup = popupGraph(mapview(breweries[1, ])@map,
                           type = "html",
                           width = 500,
                           height = 400))
# }